Make log target configurable
Right now `env_logger` always logs to stderr. This might be
undesireable in some cases.

This commit adds a way to configure the log target. It can be set to
stdout if desired.

Note that writing to stdout is implemented with `println!` instead of
`writeln!`, so that the output is properly captured in tests.