Allow custom timestamp formats.
Past experience has taught me that you'll never have one logging time
format that pleases everyone, so perhaps we should just allow it to be
configurable. :)

We can use `StrftimeItems` to parse the format string once when we build
the `Logger`. We limit the lifetime of the format string to `'static` as
it keeps things simple. (If we wanted to accept arbitrary lifetimes,
I think we'd either need to introduce a lifetime parameter on `Logger`,
or have a self-referential struct somewhere).

This commit also removes the nanoseconds, as suggested in #34. It does
not alias `+00:00` to `+Z`, as there doesn't seem to be a way to do this
with `chrono::format::strftime`. (It is possible to do it by
constructing our own iterator of formatting `Item`s, but I think
`chrono` should learn to allow this via `StftimeItems` instead).