Add ability to configure before applying env
Add `parse_env` and `parse_default_env` methods to the `Builder`, which
do the same as `from_env` and `from_default_env`, but on an existing
builder. This allows a builder to be pre-configured/obtained in another
fashion (e.g. through the [pretty_env_logger] crate), and then have the
environment applied and override the configuration.

[pretty_env_logger]: https://github.com/seanmonstar/pretty-env-logger

Signed-off-by: Fabian Schuiki <fabian@schuiki.ch>