[05 of 10]: update crate-level docs to note non-env-var configuration
For symmetry with the 'README.md' file, the crate-level API docs now
note that 'env_logger' can be configured by means other than via
environment variables, and (like README.md) directs the user to the
examples in the GitHub repo.

This is useful information for somebody first discovering the
'env_logger' crate on crates.io; it provides a more comprehensive
picture of how the crate is intended to be used.