[07 of 10]: docs: show log levels in bulleted list
The 'README.md' and crate-level API docs are both updated to show
the available logging levels in a bulleted list (rather then a
comma-separated list in the prose), and to note that they correspond
to the externally defined log::Level enum in the 'log' crate.

For the README.md file, this is an addition. The universe of valid
log levels was not previously listed there.

The order of the log levels is also now sorted in the list highest
precedence to lowest:
    error, warn, info, debug, trace

The previous order in which they were presented was:
    debug, error, info, warn, trace