use target instead of module_path in formatter by default
Target is used in filter, and its value defaults to module_path. To
enable user customize the format, also to reduce confusion, makes
formatter supporting target as well.

This change introduces a boolean builder variable "honor_target". When
set to true(the default), formatter will use target in the record
metadata.

Fix #155 and #179.